fixed / removed some warnings

+++ src/Lychee/Core/Window/Window.cpp
@@ -172,11 +172,11 @@ static uint8_t s_GLFWWindowCount = 0;
 		//		 It will not be compiled in releas builds.
 		#ifdef LY_SHOW_WINDOWTITLE_FPS
 			static u8 i = 0; 
-			static const u8 p = 10000;
+			static const u16 p = 10000;
 
 			if (i > p) {
 				i = 0;
-				u8 fps = 1 / dt.GetSeconds();
+				u8 fps = u8(1 / dt.GetSeconds());
 				std::string title =  m_sWindowData.title + " FPS:  " + std::to_string(fps);
 				glfwSetWindowTitle(m_glfwWindow, title.c_str());	
 			}
